requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
pytesseract>=0.3.10
Pillow>=10.0
pdf2image>=1.16
opencv-python-headless>=4.8